requirements.txt
# core
python-dotenv==1.0.1
pydantic==2.10.6
requests==2.32.3

# images
Pillow==10.4.0

# video/audio helpers
tqdm==4.67.1

# youtube api
google-api-python-client==2.157.0
google-auth==2.37.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
httplib2==0.22.0

# optional but recommended (better retries/backoff)
tenacity==9.0.0
